Add files via upload
In advanced technical education, laboratories are very important environment for student learning. Laboratory experience is a key factor in engineering education. In this virtual lab, student can be able to login with their credentials to the page. After completion of task, the Students can save the document with their name and Roll number. In this pandemic situation these virtual lab plays a major role , if the student can not able to attend the college at that time they can access these virtual lab through remote experimentation. Students can join into the lab with the help of internet . The page will display the problem statement , sample input and sample output. Here the task is to satisfy the test cases of the problem, it will enhance the coding skills. In this project, the students can save the documentation of programs of what they did. And they can keep the file as a reference. The status of the student can be checked by the lecturers . As well as the document taken by them also contains the name, roll number ,which is the proof that it was done by that particular student.